backend/requirements.txt
# ── API ────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── MongoDB ────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2

# ── MinIO / S3 ─────────────────────────────────────────────────
minio==7.2.7
boto3==1.34.69

# ── OCR & Image Processing ─────────────────────────────────────
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
pdfplumber==0.11.0
pypdf==4.2.0
numpy==1.26.4
scikit-image==0.23.2

# ── ML / Classification ────────────────────────────────────────
scikit-learn==1.4.2
joblib==1.4.0

# ── NLP / Extraction ───────────────────────────────────────────
spacy==3.7.4
# fr_core_news_md téléchargé via script post-install

# ── Validation ────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dateutil==2.9.0
stdnum==1.20

# ── Logging ───────────────────────────────────────────────────
structlog==24.2.0

# ── Utils ─────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
uuid==1.30
Faker==24.11.0
requests==2.31.0
